refactor(instructor): improve client handling and code modularity (#562)
