+++ instructor/__init__.py
@@ -12,14 +12,21 @@ from .dsl import (
 from .function_calls import OpenAISchema, openai_schema
 from .patch import apatch, patch
 from .process_response import handle_parallel_model
-from .client import Instructor, from_openai, from_anthropic, from_litellm, from_groq
+from .client import (
+    Instructor,
+    AsyncInstructor,
+    from_openai,
+    from_litellm,
+    Provider,
+)
+
 
 __all__ = [
     "Instructor",
     "from_openai",
-    "from_anthropic",
     "from_litellm",
-    "from_groq",
+    "AsyncInstructor",
+    "Provider",
     "OpenAISchema",
     "CitationMixin",
     "IterableModel",
@@ -36,3 +43,19 @@ __all__ = [
     "handle_parallel_model",
     "handle_response_model",
 ]
+
+try:
+    import anthropic
+    from .client_anthropic import from_anthropic
+
+    __all__.append("from_anthropic")
+except ImportError:
+    pass
+
+try:
+    import groq
+    from .client_groq import from_groq
+
+    __all__.append("from_groq")
+except ImportError:
+    pass
